Improve mouse input action prevention
* Improve mouse input action prevention

As pointed out in https://github.com/Ajatt-Tools/rikaitan/issues/2152, currently a middle mouse click still opens links in a new tab, and mouse button 4 and 5 are unusable since they trigger history back/forward.
This blocks the new tab action if the appropriate toggles are selected and adds a new set of toggles for button 4/5. I tested all of this on Chrome.

* Fix conflict with existing popup history feature

<rikaitan.link>NmJiMTRjNjY0MzRhZDc5MmI5NzhmYTk5ZDk3MmUyMjZhNGM0ZThlNwo=</rikaitan.link>